update faster rcnn example to full module api usage
