+++ example/rcnn/demo.py
@@ -2,7 +2,6 @@ import argparse
 import os
 import numpy as np
 import cv2
-import scipy.io as sio
 
 import mxnet as mx
 
@@ -10,14 +9,14 @@ from helper.processing.image_processing import resize, transform
 from helper.processing.nms import nms
 from rcnn.config import config
 from rcnn.detector import Detector
-from rcnn.symbol import get_vgg_rcnn_test
+from rcnn.symbol import get_vgg_test
 from rcnn.tester import vis_all_detection
 from utils.load_model import load_param
 
 
 def get_net(prefix, epoch, ctx):
     args, auxs = load_param(prefix, epoch, convert=True, ctx=ctx)
-    sym = get_vgg_rcnn_test()
+    sym = get_vgg_test()
     detector = Detector(sym, ctx, args, auxs)
     return detector
 
@@ -37,16 +36,14 @@ def demo_net(detector, image_name):
     :param image_name: image name
     :return: None
     """
-    # load demo data
-    im = cv2.imread(image_name + '.jpg')
-    im_array, im_scale = resize(im, config.TEST.SCALES[0], config.TRAIN.MAX_SIZE)
+    config.TEST.HAS_RPN = True
+    assert os.path.exists(image_name), image_name + ' not found'
+    im = cv2.imread(image_name)
+    im_array, im_scale = resize(im, config.SCALES[0], config.MAX_SIZE)
     im_array = transform(im_array, config.PIXEL_MEANS)
-    roi_array = sio.loadmat(image_name + '_boxes.mat')['boxes']
-    batch_index_array = np.zeros((roi_array.shape[0], 1))
-    projected_rois = roi_array * im_scale
-    roi_array = np.hstack((batch_index_array, projected_rois))
+    im_info = np.array([[im_array.shape[2], im_array.shape[3], im_scale]], dtype=np.float32)
 
-    scores, boxes = detector.im_detect(im_array, roi_array)
+    scores, boxes = detector.im_detect(im_array, im_info)
 
     all_boxes = [[] for _ in CLASSES]
     CONF_THRESH = 0.8
@@ -67,11 +64,10 @@ def demo_net(detector, image_name):
 
 
 def parse_args():
-    parser = argparse.ArgumentParser(description='Demonstrate a Fast R-CNN network')
-    parser.add_argument('--prefix', dest='prefix', help='new model prefix',
-                        default=os.path.join(os.getcwd(), 'model', 'frcnn'), type=str)
-    parser.add_argument('--epoch', dest='epoch', help='epoch of pretrained model',
-                        default=9, type=int)
+    parser = argparse.ArgumentParser(description='Demonstrate a Faster R-CNN network')
+    parser.add_argument('--image', dest='image', help='custom image', type=str)
+    parser.add_argument('--prefix', dest='prefix', help='saved model prefix', type=str)
+    parser.add_argument('--epoch', dest='epoch', help='epoch of pretrained model', type=int)
     parser.add_argument('--gpu', dest='gpu_id', help='GPU device to test with',
                         default=0, type=int)
     args = parser.parse_args()
@@ -81,5 +77,5 @@ if __name__ == '__main__':
     args = parse_args()
     ctx = mx.gpu(args.gpu_id)
     detector = get_net(args.prefix, args.epoch, ctx)
-    demo_net(detector, os.path.join(os.getcwd(), 'data', 'demo', '000004'))
-    demo_net(detector, os.path.join(os.getcwd(), 'data', 'demo', '001551'))
+    demo_net(detector, args.image)
+    demo_net(detector, args.image)
